Create make-plot.R
This script creates a figure with the monthly average rainfall (mm) in Melbourne and Oxford based off of the years 1855-2015. This script is dependent on running the combine-data.R script first to combine the two raw datasets to form the average-rainfall.csv. The final three lines of code (32-34) record the software versions used.